Admin-Module/requirements.txt
requests
django-crispy-forms